// generated/src/app/api-models/stage-auth-appeal.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class StageAuthAppeal {
  batchId: string;
  transactionId: number;
  seqNo: number;
  provId: string;
  appealNumber: string;
  appellant: string;
  contactDate: Date;
  decisionDate: Date;
  advisorDecision: string;
  notificationDate: Date;
  comments: string;
  insertDatetime: Date;
}